changes in order to user vectors to store each vector and the vectors themselves

+++ Vertice.cpp
@@ -1,8 +1,8 @@
 #include "Vertice.h"
 // Vertice::Vertice(){
-//   color = 0 ; shape = 0; weight = 0.0 ; type-0;
+//   color = 0 ; shape = 0; weight = 0 ; type-0;
 // }
-Vertice::Vertice(int c,int s, float w, int t){
+Vertice::Vertice(int c,int s, int w, int t){
   color =c; shape = s; weight = w; type = t;
 }
 void Vertice::print(){
